@@ -1,67 +1,67 @@
 {
   "name": "bkp-web",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "private": true,
+  "homepage": "https://app.blockkeeper.io",
   "dependencies": {
-    "@quentin-sommer/react-useragent": "^1.2.2",
-    "autoprefixer": "7.1.6",
-    "axios": "^0.17.0",
-    "babel-core": "6.26.0",
+    "@quentin-sommer/react-useragent": "1.2.5",
+    "autoprefixer": "8.5.0",
+    "axios": "0.18.0",
+    "babel-core": "6.26.3",
     "babel-eslint": "7.2.3",
-    "babel-jest": "20.0.3",
-    "babel-loader": "7.1.2",
-    "babel-preset-react-app": "^3.1.0",
+    "babel-jest": "23.0.0",
+    "babel-loader": "7.1.4",
+    "babel-preset-react-app": "^3.1.1",
     "babel-runtime": "6.26.0",
-    "bitcoinjs-lib": "https://github.com/yeahoffline/bitcoinjs-lib",
+    "bitcoinjs-lib": "3.3.2",
     "browser-locale": "^1.0.3",
-    "case-sensitive-paths-webpack-plugin": "2.1.1",
-    "chalk": "1.1.3",
-    "css-loader": "0.28.7",
-    "currency-symbol-map": "^4.0.3",
-    "dotenv": "4.0.0",
-    "eslint": "4.10.0",
-    "eslint-config-react-app": "^2.0.1",
-    "eslint-loader": "1.9.0",
-    "eslint-plugin-flowtype": "2.39.1",
-    "eslint-plugin-import": "2.8.0",
+    "case-sensitive-paths-webpack-plugin": "2.1.2",
+    "chalk": "2.4.1",
+    "css-loader": "0.28.11",
+    "currency-symbol-map": "^4.0.4",
+    "dotenv": "5.0.1",
+    "eslint": "4.19.1",
+    "eslint-config-react-app": "^2.1.0",
+    "eslint-loader": "2.0.0",
+    "eslint-plugin-flowtype": "2.47.1",
+    "eslint-plugin-import": "2.12.0",
     "eslint-plugin-jsx-a11y": "5.1.1",
-    "eslint-plugin-react": "7.4.0",
+    "eslint-plugin-react": "7.8.2",
     "extract-text-webpack-plugin": "3.0.2",
-    "file-loader": "1.1.5",
-    "fs-extra": "3.0.1",
-    "html-webpack-plugin": "2.29.0",
-    "jest": "20.0.4",
+    "file-loader": "1.1.11",
+    "fs-extra": "6.0.1",
+    "html-webpack-plugin": "3.2.0",
+    "jest": "23.0.0",
     "js-file-download": "^0.4.1",
-    "lzutf8": "^0.4.7",
-    "material-ui": "^1.0.0-beta.21",
-    "material-ui-icons": "^1.0.0-beta.17",
-    "moment": "^2.19.1",
+    "lzutf8": "0.5.0",
+    "@material-ui/core": "1.0.0",
+    "@material-ui/icons": "1.0.0",
+    "moment": "2.22.1",
     "object-assign": "4.1.1",
-    "postcss-flexbugs-fixes": "3.2.0",
-    "postcss-loader": "2.0.8",
+    "postcss-flexbugs-fixes": "3.3.1",
+    "postcss-loader": "2.1.5",
     "promise": "8.0.1",
     "qrcode-react": "https://github.com/yeahoffline/qrcode-react",
     "raf": "3.4.0",
-    "react": "^15.6.2",
+    "react": "^16.4.0",
     "react-copy-to-clipboard": "^5.0.1",
-    "react-cryptocoins": "^1.0.9",
-    "react-dev-utils": "^4.2.1",
-    "react-dom": "^15.6.2",
-    "react-qr-reader": "^2.0.0",
+    "react-cryptocoins": "^1.0.11",
+    "react-dev-utils": "^5.0.1",
+    "react-dom": "^16.4.0",
+    "react-qr-reader": "^2.1.0",
     "react-router-dom": "^4.2.2",
     "react-scroll-up": "^1.3.3",
-    "react-transitive-number": "^3.1.0",
-    "style-loader": "0.19.0",
-    "sw-precache-webpack-plugin": "0.11.4",
-    "url-loader": "0.6.2",
-    "uuid": "^3.1.0",
-    "validator": "^9.1.1",
-    "web3-utils": "^1.0.0-beta.24",
+    "style-loader": "0.21.0",
+    "sw-precache-webpack-plugin": "0.11.5",
+    "url-loader": "1.0.1",
+    "uuid": "^3.2.1",
+    "validator": "^10.2.0",
+    "web3-utils": "^1.0.0-beta.34",
     "webpack": "3.8.1",
     "webpack-dev-server": "2.9.4",
     "webpack-manifest-plugin": "1.3.2",
-    "whatwg-fetch": "2.0.3",
-    "yarn": "^1.3.2"
+    "whatwg-fetch": "2.0.4",
+    "yarn": "^1.7.0"
   },
   "scripts": {
     "start": "node scripts/start.js",
