@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import registerServiceWorker from './registerServiceWorker'
 import {BrowserRouter, Route, Switch, Redirect} from 'react-router-dom'
-import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider'
+import MuiThemeProvider from '@material-ui/core/styles/MuiThemeProvider'
 import {theme} from './view/Style'
 import Core from './logic/Core'
 import UserView from './view/User'
